Remove failed higher-level version using gzip module
This commit is not squashed because the gzip version might serve a
purpose later. As described in the original commit message, one might
attempt to recover more data by monkey-patching the _read_eof()
function. For now, it's too awkward to fix up the state of the
decompressor from a high level and the low-level approach works quite
nicely.