DEV-1164 Script to create API files for RNA
Add file entries in the API for the RNA FASTQ files. These are not
registered with the API (contrast with the pipeline output which has
this done at the end of its run) when they are created so this script
can be called periodically to do it.

Note that because the RNA is not part of a run, a synthetic run has been
used.