DEV-1351 Make reference arrivals an event
Initially the reference arrivals were stored against their own reference
sample. This has been changed so that there are now events for each of
the tumour and reference, but only one sample (named after the tumour).

Additionally the reported events have been improved to include the
report type which is quite informative in non-success cases.