DEV-1843 Split "to" and "from" sync for resources
We need to have the samplesheets and LIMS data copied to GCP as soon as
possible after it becomes available.  Logs and JSONs resulting from the
registration operations however are not that critical. Split them up and
then they can be done on appropriate schedules as the majority of the
time taken up until now has been synching the registration output back
to the datastore machine.