DEV-2208 Sync script for data-vm resources
After some investigation it seems the least complex way to expose the correct data from ops to data vm
is to have a specific script which lays out the files exactly the are in ops in the bucket. The alternative
would be lots of switches in the scripts to have the locate_ functions work in multiple contexts.

The script copies all files defined in DEV-2208 into a bucket used just for this purpose, which will be
synced back down to the data vm