DEV-2369 Switch to pipeline versioning
The pipeline tools (in `gs://common-tools`) use the path scheme
`/${toolname}/${version}/${toolname}.jar`, while the `datastore`-era path scheme
was `/${toolname}_v${version}/${toolname}.jar`. I've changed most of
those in this changest but did leave a couple in the `bam` subdirectory
because I'm not sure if they're going to be used in GCP VMs.

Also I've removed the `common` from `/data/common/...` so that both the
VMs we currently have are the same.