Fix bug due to incorrect syntax for filtering in jq
Instead of filtering a list of results from the API by the value of one
of the fields in each result, the code was selecting all lists of
results such that at least one result in the list had
the desired value. So what we wanted was that it only kept those
results with the desired values. What it did was that if at least one
of the results in the list had the desired value then it would keep
the entire list, and if none of the results had the desired value then it
would keep nothing.