DEV-2438 Avoid syncing LIMS logs
We are having intermittent problems to do with the log being updated
long after expected and causing issues with the copy up to the bucket.

Just copy the TSVs and the lims.json, nothing else should really be
required, seems like sometimes are are temporary backups made and that
sort of thing but should not be required.