DEV-2480 Stripped-down uploader for BCLs
Simplify the sync to GCP down to bare-minimum.

Most of the functionality of the upload script will now live on the GCP
side; this script now only really does the "rsync" to the bucket. It
does have one new responsibility though which is to write the timestamp
of the local RTAComplete.txt file into another file so it survives the
journey to GCP intact.

A service file has also been added to allow the systemd on the host
system to keep this script running but also guarantee a single copy only
gets started, which seems a more natural fit than `cron`. This
essentially means this script will constantly be copying data from
flowcells as they are streaming up to the host (currently the crunches).

On the GCP side the code will detect when a flowcell is complete based
on its contents. There is likely to be some more code later to help with
coordinating the deletion of data after it has been confirmed to be
present in GCP.

Note that this new approach should help improve the overall processing
time as sequencer output will be eligible for copying up to the bucket
immediately rather than having to wait for the sequencer to finish
writing the run they belong to.