DEV-2480 Anticipate bug in rsyncing
Exclude the `RTAComplete.txt` file from the `gsutil rsync`, and do a
second `gsutil rsync` after it's done, but only if the file then exists,
and only then copy the `RTAComplete.txt` file.

The reason for the histrionics is that we're going to be synching files
up as they become available, not only once `RTAComplete.txt` is written.
Because the GCP-side component will look for that file to determine
whether a flowcell is ready for processing _and_ we don't control the
order in which files are `rsync`ed, we need to hold that file back until
we're sure everything else has been synchronised for the flowcell, only
pushing up the text file once everything else has for sure been copied.

Also introduce an error handler function and a `set -e` at the start of
the script because we can't afford to have any function run after one
rsync fails, and exiting a script when a function fails in BASH is
tricky business.