Move calculation for extra yield after shallow to Python script
The calculation in bash was using some tricks to do
fixed-point arithmetic in bash. This trick had a bug for samples
where excl starts with 0.0, and also when excl==0. Instead of
fixing this bug in the trick, I decided to prevent future bugs by
removing the trick and handling the calculation in an environment
that can simply handle this calculation properly.