DEV-2578 Wait 20m before marking flowcell done
We had a problem the other day where the last two BCLs for a flowcell
didn't get automatically copied up and it appears to be a timing issue
between when the `RTAComplete.txt` file was written and when the BCLs
were written. They were within 10m of each other based on the logs but
the `gsutil rsync` did not detect the files so it seems they were not
present when it started constructing its synchronisation list.

Incidentally I believe this has been aggravated by our change of last
week to stop trying to synchronise flowcells that we already consider
"done". I think in the past if this happened at all the next sync would
get the file and the problem wouldn't appear. At least we know the
`bcl2fastq` process is not just blindly processing the data with no
checks.

Switch to using `find` with the `mmin` query to only do something after
the `RTAComplete.txt` file has existed for at least 20 minutes. The
implementation is a bit unkempt because `find` does not return non-zero
when it finds no files, only if it hits an error. So we have to
interpret the results with a `grep .`, which _will_ return non-zero if
there is no output.