DEV-2578 Remove error logging
In practice this has resulted in many false alarms since the only thing
we regularly see is a failure resulting from the directory contents
changing between the time when `gsutil rsync` starts and when it has
completed constructing its list of files to synchronise.

Unfortunately there doesn't appear to be any way around this currently
as the files that are causing the error are in the list of exclusions.
However the problem is that the files aren't ignored during the
construction of the list.

For now this is a calculated risk and we feel that suppressing the
errors is safe enough given the GCP-side processing does appear to be
smart enough to refuse to process incorrect or incomplete data.