DEV-2564 Support new verification flow
With this approach we will have a new DB schema every time we do a
verification, so we will no longer have to rename samples for every
run.

The changes here add the arguments we'll need to allow the use of two
schemas, as the previous flow assumed a single database with different
sample names for each run.

There is probably more to do here, I haven't tested everything yet.