OPR-113 Modify logo for GCP patient-reporter
The application currently hardcodes the path of the file in the bucket
to use for the logo in the report. I don't see a significant advantage
to changing the code of the application to accept an alternative path
for this image, so have just modified the sync script to overwrite the
expected file with the one containing the logo we now want to use.