ACTIN-1102 Scripts for sharing NKI reports
The pipeline will no longer name the reports by the NKI hashed id,
writing them instead under the ACTIN id to simplify our operational
processes.

These new scripts will automate the process of copying the reports to
the share bucket under the name expected by Kaiko, and also expose the
ability to lookup the hashed id for a given ACTIN id.

I am aware that these do not utilise the "find..." functions, I am not
sure they are warranted in this case as I do not think they will be used
beyond NKI for the moment. That may come as part of a later update.