INFRA-755 fix clinical and DR view with missing treatments
When a biopsy is not defined for a sample, but a treatment is, the
clinical and datarequest views do not join the treatment because they
join on the biopsy ID. This is the case for 239 CPCT samples and 138
WIDE samples. The fix is to join on the patient ID instead of the biopsy
ID. The tricky thing about it is that we only want to join on the first
treatment, but after the first biopsy, if it exists. To achieve this, I
created the first treatment after biopsy view.

This workaround can be removed once the clinical biopsies are reworked
and the data from LAMA is incorporated.